Extract parent class `LikedUsersListPopup`.
