+++ tests/photo_page_elements.py
@@ -10,8 +10,6 @@ from selenium.webdriver import ActionChains
 from page_elements import PageElement, LikesController, LikeButton, UnlikeButton, \
     LikedUsersListPopup
 
-from utils import custom_move_to_element
-
 
 class LikePhotoUnderPhoto(LikeButton):
 
@@ -48,17 +46,6 @@ class PhotoLikedUsers(LikedUsersListPopup):
     # LINK_CSS_SELECTOR = 'div.photo-layer_bottom_block ul.widget-list > li:last() > div:nth-child(1) button'
     LINK_XPATH = '//div[contains(@class, "photo-layer_bottom_block __actions")]//descendant::li[last()]/descendant::button'
     USERNAME_LINK = '(//ul[@class="ucard-mini-list"]/li/descendant::div[@class="ucard-mini_cnt_i ellip"])[1]'
-
-    def has_your_like(self, username):
-        username = unicode(username, 'utf8')
-        custom_move_to_element(
-            self.driver, self.LINK_XPATH, click_times=2
-        )
-
-        first_user = WebDriverWait(self.driver, 30).until(
-            EC.visibility_of_element_located((By.XPATH, self.USERNAME_LINK))
-        )
-        return first_user.text == username
 
 
 class PhotoLikedUsersInAlbum(PhotoLikedUsers):
