+++ tests/video_page_elements.py
@@ -13,54 +13,69 @@ from page_elements import PageElement, LikesController, LikeButton, UnlikeButton
 from utils import custom_move_to_element
 
 
-class LikePhotoUnderPhoto(LikeButton):
+class LikeVideoButtonUnderPlayer(LikeButton):
 
-    BUTTON = '//div[contains(@class, "photo-layer_bottom_block __actions")]//descendant::li[last()]'
-
-
-class UnlikePhotoUnderPhoto(UnlikeButton):
-
-    BUTTON = '//div[contains(@class, "photo-layer_bottom_block __actions")]//descendant::li[last()]'
+    BUTTON = '//ul[@class="widget-list"]/descendant::button[last()]'
 
 
-class LikePhotoInTheRightPhotoCorner(LikeButton):
+class UnlikeVideoButtonUnderPlayer(UnlikeButton):
 
-    BUTTON = '//span[@class="ico_big-klass-white"]'
-
-
-class UnlikePhotoInTheRightPhotoCorner(UnlikeButton):
-
-    BUTTON = '//span[@class="ico_big-klass-white"]'
+    BUTTON = '//ul[@class="widget-list"]/descendant::button[last()]'
 
 
-class LikeNotOpenedPhotoInAlbum(LikeButton):
+# class LikePhotoInTheRightPhotoCorner(LikeButton):
 
-    BUTTON = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
-
-
-class UnlikeNotOpenedPhotoInAlbum(UnlikeButton):
-
-    BUTTON = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
+#     BUTTON = '//span[@class="ico_big-klass-white"]'
 
 
-class PhotoLikedUsers(LikedUsersListPopup):
+# class UnlikePhotoInTheRightPhotoCorner(UnlikeButton):
 
-    # LINK_CSS_SELECTOR = 'div.photo-layer_bottom_block ul.widget-list > li:last() > div:nth-child(1) button'
-    LINK_XPATH = '//div[contains(@class, "photo-layer_bottom_block __actions")]//descendant::li[last()]/descendant::button'
-    USERNAME_LINK = '(//ul[@class="ucard-mini-list"]/li/descendant::div[@class="ucard-mini_cnt_i ellip"])[1]'
+#     BUTTON = '//span[@class="ico_big-klass-white"]'
 
-    def has_your_like(self, username):
-        username = unicode(username, 'utf8')
-        custom_move_to_element(
-            self.driver, self.LINK_XPATH, click_times=2
+
+# class LikeNotOpenedPhotoInAlbum(LikeButton):
+
+#     BUTTON = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
+
+
+# class UnlikeNotOpenedPhotoInAlbum(UnlikeButton):
+
+#     BUTTON = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
+
+
+# class PhotoLikedUsers(LikedUsersListPopup):
+
+#     # LINK_CSS_SELECTOR = 'div.photo-layer_bottom_block ul.widget-list > li:last() > div:nth-child(1) button'
+#     LINK_XPATH = '//div[contains(@class, "photo-layer_bottom_block __actions")]//descendant::li[last()]/descendant::button'
+#     USERNAME_LINK = '(//ul[@class="ucard-mini-list"]/li/descendant::div[@class="ucard-mini_cnt_i ellip"])[1]'
+
+#     def has_your_like(self, username):
+#         username = unicode(username, 'utf8')
+#         custom_move_to_element(
+#             self.driver, self.LINK_XPATH, click_times=2
+#         )
+
+#         first_user = WebDriverWait(self.driver, 30).until(
+#             EC.visibility_of_element_located((By.XPATH, self.USERNAME_LINK))
+#         )
+#         return first_user.text == username
+
+
+class VideoLikedUsers(LikedUsersListPopup):
+
+    LINK = '//ul[@class="ucard-mini-list"]/li/descendant::div[@class="ucard-mini_cnt_i ellip"]'
+
+    def has_your_like(self, login):
+        xpath = '//ul[@class="widget-list"]/descendant::button[last()]'
+        custom_move_to_element(self.driver, xpath)
+
+        login = unicode(login, 'utf8')
+        first_liked_user_in_popup = WebDriverWait(self.driver, 30, 0.1).until(
+            EC.visibility_of_element_located((By.XPATH, self.LINK))
         )
-
-        first_user = WebDriverWait(self.driver, 30).until(
-            EC.visibility_of_element_located((By.XPATH, self.USERNAME_LINK))
-        )
-        return first_user.text == username
+        return first_liked_user_in_popup.text == login
 
 
-class PhotoLikedUsersInAlbum(PhotoLikedUsers):
+# class PhotoLikedUsersInAlbum(PhotoLikedUsers):
 
-    LINK_XPATH = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
+#     LINK_XPATH = '(//li[contains(@class, "ugrid_i")])[1]//descendant::li[1]'
